framework-less/Live-API/RAG-Engine/text2text/requirements.txt
google-generativeai
google-cloud-aiplatform
pydantic
uvicorn
fastapi